Add shared mock for core/editor store in tests
Create a reusable test utility for mocking the core/editor store.
This is needed because we use string-based store selectors instead of
importing the store from @wordpress/editor (to avoid wp-editor dependency).
