+++ plugins/woocommerce/client/blocks/assets/js/blocks/checkout/test/block.tsx
@@ -43,23 +43,10 @@ import Taxes from '../inner-blocks/checkout-order-summary-taxes/frontend';
 import { defaultCartState } from '../../../data/cart/default-state';
 import Checkout from '../block';
 
-jest.mock( '@wordpress/data', () => {
-	const wpData = jest.requireActual( 'wordpress-data-wp-6-7' );
-	// Create a mock editor store with basic selectors
-	const mockEditorStore = wpData.createReduxStore( 'core/editor', {
-		reducer: () => ( {} ),
-		selectors: {
-			getCurrentPostId: () => 0,
-			getCurrentPostType: () => 'post',
-			isCurrentPostPublished: () => false,
-		},
-	} );
-	wpData.register( mockEditorStore );
-	return {
-		__esModule: true,
-		...wpData,
-	};
-} );
+jest.mock( '@wordpress/data', () =>
+	// eslint-disable-next-line @typescript-eslint/no-var-requires -- Must use require due to Jest mock hoisting
+	require( '@woocommerce/blocks-test-utils/mock-editor-store' ).mockWordPressDataWithEditorStore()
+);
 
 jest.mock( '@wordpress/compose', () => ( {
 	...jest.requireActual( '@wordpress/compose' ),
